+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Upgrade to @mangrovedao/mangrove-deployments v2.0.1
 - Upgrade to @mangrovedao/context-addresses v1.1.0
 - Simplify copying of context addresses
+- Add support for Blast Sepolia
 
 # 2.0.3